+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/generators/string/CommentGenerator.kt
@@ -1,8 +1,0 @@
-package com.windea.breezeframework.data.generators.string
-
-import com.windea.breezeframework.data.generators.*
-
-/**注释生成器。*/
-object CommentGenerator : Generator {
-	//TODO
-}